Changes based on review comments
